backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
